Add support for mgmt vrf
When management vrf is enabled and vrf is present in the tacacs config,
if we are unable to reach any configured tacacs server, try setting
vrf context on the socket.

Previously libnss-tacplus worked only with ssh@mgmt, now works with normal
ssh in mgmt vrf

Setting via the socket (rather than vrf context) is required so we don't
set the VRF context for arbitrary processes that do uid or username
lookups.