During login from ssh, send remote host IP address in AUTH request
The hack is to run getpeername on fd 0, because during ssh connections,
it is a socket from the remote host.  This is a bit fragile...

Normally fd 0 interactively will be a pty or tty, so getpeername() will fail.

There may be some daemons where fd0 is a socket, and returns a local or
some other remote IP address, and if so, it could lead to some
confusion, but it shouldn't ever break anything.

I ran with tshark watching the packet exchange, and verified that the
remote address field is set for ssh sessions at the start of the ssh
session, and not when run in other uses.  The customer ran a 3.2.1
package with this change, and it resolved their issue.