Track changes to config files, and reparse if any change
This is done to handle the case where nss_tacplus.so is included in
a long-lived daemon.  It's desirable to have long-lived daemons reflect
changes to the configuration, both to enable/disable debugging, and
particularly if the server list or key changes.  Clear all read config
variables to defaults when re-parsing.

This is complicated by nested configuration files via the include
directive.   At top level, we need to check all the previously used
configuration files to see if any have changed.

This also adds a limitation to no more than 8 deep include nesting.  In
practice, > 2 is going to be very rare, so it should be OK.

Log a message when we re-initialize (without using debug qualifier).